feat: progressbar on bsp load
Adds a progressbar (that scrolls LTR) so that the user knows something is happening while the bsp is being loaded. Like Tom mentioned, a percentage based progressbar is kind of meaningless since it'll go to around 99% in an instant and the last 1% is what will take up all the time.